+++ modules/sdk-coin-icp/src/lib/utils.ts
@@ -1,12 +1,10 @@
-import { BaseUtils } from '@bitgo/sdk-core';
-import elliptic from 'elliptic';
+import { BaseUtils, KeyPair } from '@bitgo/sdk-core';
+import { secp256k1 } from '@noble/curves/secp256k1';
 import { Principal as DfinityPrincipal } from '@dfinity/principal';
 import * as agent from '@dfinity/agent';
 import crypto from 'crypto';
 import crc32 from 'crc-32';
 import { KeyPair as IcpKeyPair } from './keyPair';
-
-const Secp256k1Curve = new elliptic.ec('secp256k1');
 
 export class Utils implements BaseUtils {
   isValidAddress(address: string): boolean {
@@ -17,18 +15,47 @@ export class Utils implements BaseUtils {
     throw new Error('Method not implemented.');
   }
 
-  isValidPublicKey(key: string): boolean {
-    const hexRegex = /^[0-9a-fA-F]+$/;
-    if (!hexRegex.test(key)) return false;
+  isValidPublicKey(hexStr: string): boolean {
+    if (!this.isValidHex(hexStr)) {
+      return false;
+    }
 
-    const length = key.length;
-    if (length !== 130) return false;
+    if (!this.isValidLength(hexStr)) {
+      return false;
+    }
 
-    return true;
+    const pubKeyBytes = this.hexToBytes(hexStr);
+    const firstByte = pubKeyBytes[0];
+    return (
+      (pubKeyBytes.length === 33 && (firstByte === 2 || firstByte === 3)) ||
+      (pubKeyBytes.length === 65 && firstByte === 4)
+    );
   }
 
+  isValidLength(hexStr: string): boolean {
+    return hexStr.length / 2 === 33 || hexStr.length / 2 === 65;
+  }
+
+  isValidHex(hexStr: string): boolean {
+    return /^([0-9a-fA-F]{2})+$/.test(hexStr);
+  }
+
+  hexToBytes(hex: string): Uint8Array {
+    return new Uint8Array(Buffer.from(hex, 'hex'));
+  }
+
+  /** @inheritdoc */
   isValidPrivateKey(key: string): boolean {
-    throw new Error('Method not implemented.');
+    return this.isValidKey(key);
+  }
+
+  isValidKey(key: string): boolean {
+    try {
+      new IcpKeyPair({ prv: key });
+      return true;
+    } catch {
+      return false;
+    }
   }
 
   isValidSignature(signature: string): boolean {
@@ -53,13 +80,18 @@ export class Utils implements BaseUtils {
   }
 
   compressPublicKey(uncompressedKey: string): string {
-    if (!uncompressedKey.startsWith('04')) {
-      throw new Error('Invalid uncompressed public key format');
+    if (uncompressedKey.startsWith('02') || uncompressedKey.startsWith('03')) {
+      return uncompressedKey;
     }
+    if (!uncompressedKey.startsWith('04') || uncompressedKey.length !== 130) {
+      throw new Error('Invalid uncompressed public key format.');
+    }
+
     const xHex = uncompressedKey.slice(2, 66);
     const yHex = uncompressedKey.slice(66);
     const y = BigInt(`0x${yHex}`);
     const prefix = y % 2n === 0n ? '02' : '03';
+
     return prefix + xHex;
   }
 
@@ -68,11 +100,9 @@ export class Utils implements BaseUtils {
   }
 
   derivePrincipalFromPublicKey(publicKeyHex: string): DfinityPrincipal {
-    const publicKeyBuffer = Buffer.from(publicKeyHex, 'hex');
-
     try {
-      const ellipticKey = Secp256k1Curve.keyFromPublic(publicKeyBuffer);
-      const uncompressedPublicKeyHex = ellipticKey.getPublic(false, 'hex');
+      const point = secp256k1.ProjectivePoint.fromHex(publicKeyHex);
+      const uncompressedPublicKeyHex = point.toHex(false);
       const derEncodedKey = agent.wrapDER(Buffer.from(uncompressedPublicKeyHex, 'hex'), agent.SECP256K1_OID);
       const principalId = DfinityPrincipal.selfAuthenticating(Buffer.from(derEncodedKey));
       const principal = DfinityPrincipal.fromUint8Array(principalId.toUint8Array());
@@ -107,6 +137,18 @@ export class Utils implements BaseUtils {
     const compressedKey = this.compressPublicKey(hexEncodedPublicKey);
     const KeyPair = new IcpKeyPair({ pub: compressedKey });
     return KeyPair.getAddress();
+  }
+
+  public generateKeyPair(seed?: Buffer): KeyPair {
+    const keyPair = seed ? new IcpKeyPair({ seed }) : new IcpKeyPair();
+    const keys = keyPair.getKeys();
+    if (!keys.prv) {
+      throw new Error('Missing prv in key generation.');
+    }
+    return {
+      pub: keys.pub,
+      prv: keys.prv,
+    };
   }
 }