+++ modules/sdk-core/src/bitgo/lightning/index.ts
@@ -1,5 +1,0 @@
-export * from './signableJson';
-export * from './signature';
-export * from './lightningUtils';
-export * from './lightningWallet';
-export * from './codecs';