+++ modules/statics/src/tokenConfig.ts
@@ -44,13 +44,17 @@ export interface BaseContractAddressConfig extends BaseNetworkConfig {
 export type AvaxcTokenConfig = BaseContractAddressConfig;
 export type CeloTokenConfig = BaseContractAddressConfig;
 export type EthLikeTokenConfig = BaseContractAddressConfig;
-export type EosTokenConfig = BaseContractAddressConfig;
+export type EosTokenConfig = BaseContractAddressConfig & {
+  contractName: string;
+  contractAddress: string;
+};
 export type Erc20TokenConfig = BaseContractAddressConfig;
 export type TrxTokenConfig = BaseContractAddressConfig;
 export type StellarTokenConfig = BaseNetworkConfig;
 
 export type SolTokenConfig = BaseNetworkConfig & {
   tokenAddress: string;
+  contractAddress: string;
 };
 
 export type AdaTokenConfig = BaseNetworkConfig & {
@@ -328,6 +332,8 @@ const formattedEosTokens = coins.reduce((acc: EosTokenConfig[], coin) => {
       name: coin.fullName,
       tokenContractAddress: coin.contractName.toString().toLowerCase(),
       decimalPlaces: coin.decimalPlaces,
+      contractName: coin.contractName,
+      contractAddress: coin.contractAddress,
     });
   }
   return acc;
@@ -429,6 +435,7 @@ const formattedSolTokens = coins.reduce((acc: SolTokenConfig[], coin) => {
       name: coin.fullName,
       tokenAddress: coin.tokenAddress,
       decimalPlaces: coin.decimalPlaces,
+      contractAddress: coin.contractAddress,
     });
   }
   return acc;