+++ modules/express/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [11.0.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/express@10.7.0...@bitgo/express@11.0.0) (2025-02-20)
+
+### Bug Fixes
+
+- **express:** correct parsing of initwallet ([9e9601d](https://github.com/BitGo/BitGoJS/commit/9e9601d3bdd5f03a9177babbac47a63360c5a546))
+
+### Features
+
+- **abstract-lightning:** move lnv2 logic from sdk-core ([5c9114d](https://github.com/BitGo/BitGoJS/commit/5c9114d6a4dbdd5a130bde7cb897ff6f1f8132e0))
+- **express:** move walletId to the path ([1c6b555](https://github.com/BitGo/BitGoJS/commit/1c6b55575c5e75b1292a73f346a8d2db658743d0))
+
+### BREAKING CHANGES
+
+- **express:** changed path for lightning apis
+
+TICKET: BTC-1846
+
+- **abstract-lightning:** Lightning v2
+
+# [10.7.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/express@10.6.14...@bitgo/express@10.7.0) (2025-02-19)
+
+### Features
+
+- make expressip macaroon caveat optional ([1d2bd03](https://github.com/BitGo/BitGoJS/commit/1d2bd033073209fea1353bea44d55ad5078296b4))
+- **sdk-core:** move lightning specific wallet functions ([e63129d](https://github.com/BitGo/BitGoJS/commit/e63129dfe0e910ed44fe80bdf42dba5edc7e76a5))
+
 ## [10.6.14](https://github.com/BitGo/BitGoJS/compare/@bitgo/express@10.6.13...@bitgo/express@10.6.14) (2025-02-11)
 
 **Note:** Version bump only for package @bitgo/express