+++ modules/sdk-coin-eos/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.60](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-eos@2.1.59...@bitgo/sdk-coin-eos@2.1.60) (2025-02-20)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-eos
+
+## [2.1.59](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-eos@2.1.58...@bitgo/sdk-coin-eos@2.1.59) (2025-02-19)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-eos
+
 ## [2.1.58](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-eos@2.1.57...@bitgo/sdk-coin-eos@2.1.58) (2025-02-11)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-eos